+++ packages/compat/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.20](https://github.com/react-navigation/react-navigation/compare/@react-navigation/compat@5.3.19...@react-navigation/compat@5.3.20) (2021-08-17)
+
+
+### Bug Fixes
+
+* make compat layer work with 6.x ([c9c42c6](https://github.com/react-navigation/react-navigation/commit/c9c42c62d553ed3ceb64de59227a065763087728))
+
+
+
+
+
+## [5.3.19](https://github.com/react-navigation/react-navigation/compare/@react-navigation/compat@5.3.18...@react-navigation/compat@5.3.19) (2021-08-17)
+
+**Note:** Version bump only for package @react-navigation/compat
+
+
+
+
+
+## [5.3.18](https://github.com/react-navigation/react-navigation/compare/@react-navigation/compat@5.3.17...@react-navigation/compat@5.3.18) (2021-08-15)
+
+**Note:** Version bump only for package @react-navigation/compat
+
+
+
+
+
+## [5.3.17](https://github.com/react-navigation/react-navigation/compare/@react-navigation/compat@5.3.15...@react-navigation/compat@5.3.17) (2021-08-01)
+
+**Note:** Version bump only for package @react-navigation/compat
+
+
+
+
+
+## [5.3.16](https://github.com/react-navigation/react-navigation/compare/@react-navigation/compat@5.3.15...@react-navigation/compat@5.3.16) (2021-08-01)
+
+**Note:** Version bump only for package @react-navigation/compat
+
+
+
+
+
+## [5.3.15](https://github.com/react-navigation/react-navigation/compare/@react-navigation/compat@5.3.14...@react-navigation/compat@5.3.15) (2021-04-04)
+
+**Note:** Version bump only for package @react-navigation/compat
+
+
+
+
+
 ## [5.3.14](https://github.com/react-navigation/react-navigation/compare/@react-navigation/compat@5.3.13...@react-navigation/compat@5.3.14) (2021-02-21)
 
 **Note:** Version bump only for package @react-navigation/compat