+++ packages/material-top-tabs/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.19](https://github.com/react-navigation/react-navigation/compare/@react-navigation/material-top-tabs@5.3.18...@react-navigation/material-top-tabs@5.3.19) (2021-08-17)
+
+**Note:** Version bump only for package @react-navigation/material-top-tabs
+
+
+
+
+
+## [5.3.18](https://github.com/react-navigation/react-navigation/compare/@react-navigation/material-top-tabs@5.3.17...@react-navigation/material-top-tabs@5.3.18) (2021-08-15)
+
+**Note:** Version bump only for package @react-navigation/material-top-tabs
+
+
+
+
+
+## [5.3.17](https://github.com/react-navigation/react-navigation/compare/@react-navigation/material-top-tabs@5.3.15...@react-navigation/material-top-tabs@5.3.17) (2021-08-01)
+
+**Note:** Version bump only for package @react-navigation/material-top-tabs
+
+
+
+
+
+## [5.3.16](https://github.com/react-navigation/react-navigation/compare/@react-navigation/material-top-tabs@5.3.15...@react-navigation/material-top-tabs@5.3.16) (2021-08-01)
+
+**Note:** Version bump only for package @react-navigation/material-top-tabs
+
+
+
+
+
+## [5.3.15](https://github.com/react-navigation/react-navigation/compare/@react-navigation/material-top-tabs@5.3.14...@react-navigation/material-top-tabs@5.3.15) (2021-04-04)
+
+**Note:** Version bump only for package @react-navigation/material-top-tabs
+
+
+
+
+
 ## [5.3.14](https://github.com/react-navigation/react-navigation/compare/@react-navigation/material-top-tabs@5.3.13...@react-navigation/material-top-tabs@5.3.14) (2021-02-21)
 
 **Note:** Version bump only for package @react-navigation/material-top-tabs